pos update fixes

+++ web/public/outputs.js
@@ -1,18 +1,16 @@
 const PAYLOAD_ADDR = "http://0.0.0.0:7070";
 
-
-// Signals to the actuator
-// long positionCommand; // (range: -1000 to 1000)
-// long forceCommand;  // (range: 0 to 1023)
-// bool activated; // Not used
-// bool airOut;  // Set high to open air-out valve
-// bool airIn;   // Set high to open air-in valve
-
 export function sendPositionUpdate(positionValue, forceValue, airOut, airIn) {
+    // Signals to the actuator
+    // long positionCommand; // (range: -1000 to 1000)
+    // long forceCommand;  // (range: 0 to 1023)
+    // bool activated; // Not used
+    // bool airOut;  // Set high to open air-out valve
+    // bool airIn;   // Set high to open air-in valve
     const buffer = new ArrayBuffer(3);
     const view = new DataView(buffer);
-    view.setUint8(0, positionValue);
-    view.setUint8(1, forceValue);
+    view.setInt16(0, positionValue);
+    view.setInt16(1, forceValue);
 
     let boolArr = 0;
     if (airOut) {
